requirements.txt
faiss-cpu
sentence-transformers
requests
gradio
numpy
scikit-learn
torch
python-docx